OurCNN/requirements.txt
# Core ML and Deep Learning
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.13.0
keras>=2.13.0

# Computer Vision and Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0

# YOLO and Object Detection
ultralytics>=8.0.0
yolov5>=7.0.0

# Data Processing and Analysis
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Signal Processing (for RF signals)
librosa>=0.10.0
scipy>=1.11.0

# Data Loading and Augmentation
albumentations>=1.3.0
imgaug>=0.4.0

# Progress and Utilities
tqdm>=4.65.0
wandb>=0.15.0
tensorboard>=2.13.0

# Configuration and File Handling
pyyaml>=6.0
toml>=0.10.0
h5py>=3.9.0

# Jupyter and Development
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=6.5.0

# Additional ML Utilities
xgboost>=1.7.0
lightgbm>=4.0.0
optuna>=3.3.0

# Audio/Signal Processing (for spectrograms)
librosa>=0.10.0
soundfile>=0.12.0

# Memory and Performance
psutil>=5.9.0
memory-profiler>=0.60.0

# Testing and Validation
pytest>=7.4.0
pytest-cov>=4.1.0
